dashboard/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings==2.6.1
aiosqlite==0.20.0
httpx==0.27.2
oracledb>=2.3.0
APScheduler==3.10.4
openpyxl==3.1.5
python-multipart==0.0.9
